client uses integers for race times now
Conflicts:
	src/game/client/components/hud.cpp
	src/game/client/components/menus_ingame.cpp
	src/game/client/components/race_demo.cpp
	src/game/client/components/scoreboard.cpp
	src/game/client/gameclient.cpp

+++ src/game/client/components/hud.h
@@ -15,8 +15,8 @@ class CHud : public CComponent
 	int m_LastReceivedTimeTick;
 	int m_CheckpointTick;
 	int m_RaceTick;
-	float m_Record;
-	float m_FinishTime;
+	int m_Record;
+	int m_FinishTime;
 	int m_RaceState;
 
 	enum
