+++ src/game/client/components/race_demo.h
@@ -9,7 +9,7 @@ class CRaceDemo : public CComponent
 {
 	int m_RecordStopTime;
 	int m_DemoStartTick;
-	float m_Time;
+	int m_Time;
 	const char *m_pMap;
 	
 	bool m_Active;
